Reconciled against the trunk.
git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/branches/cairo_32bitColor@12718 10fe2890-8a40-4696-8b1d-7ccbb19c398e

+++ ni/src/lib/hlu/NresDB.h
@@ -427,7 +427,7 @@ extern NhlBoolean NrmQinQList(
 #endif
 );
 
-typedef unsigned long Signature;
+typedef unsigned int Signature;
 
 extern NrmQuark _NrmInternalStringToQuark(
 #if	NhlNeedProto
